function/requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-texttospeech
